got API records going into db
+ FINALY got the GraphQL API response working...i didnt have the right URL :|
+ Fixed up the database with sqlalchamy, now all records are being entered in properly!

Now an algorithm needs to be developed next to ban / challange etc.
thoughts: 100 request a min will result in a JS challange, 160 will result in a CAPTCHA challange, 300 in a minute will be a ban.

Then need to implement the API to take action! then ont to actually developing DDN